vmwgfx: Fix a hibernation error path
If fifo resources for some reason are active when we hibernate
(which is an error in itself), the hibernation code refuses to hibernate and
enters an error path. A recent code rewrite caused a regression that caused the
device not to be setup properly when recovering from the initial stages of
hibernation.

Signed-off-by: Thomas Hellstrom <thellstrom@vmware.com>
Reviewed-by: Sinclair Yeh <syeh@vmware.com>