vmwgfx: Add temporary support for drm_ioctl_mode_cursor2
Newer desktop environments have switched over to using the new
cursor2 ioctl.  Porting this capability over is not straightforward
given the current state of the stand alone driver.

Since we do not support cursor hotspot, an easier solution is to
simply throw away hot_x and hot_y, and call the existing cursor
ioctl.

We will give this a proper update when we update everything for
the atomic mode set feature.

Signed-off-by: Sinclair Yeh <syeh@vmware.com>
Reviewed-by: Thomas Hellstrom <thellstrom@vmware.com>