drm: Implement calling the cursor_set2 driver callback
So that the new compositors can communicate the hotspot to us.
This was previously done using a driver-specific ioctl.

The version upstream is setting the hotspot to (0,0) if the user-space
client is using the drm_mode_cursor_ioctl(). We can't do that as that
would reset a hotspot previously set using our driver-private ioctl,
so we need to propagate that change also upstream before implementing
cursor_set2.

Signed-off-by: Thomas Hellstrom <thellstrom@vmware.com>
Reviewed-by: Sinclair Yeh <syeh@vmware.com>