vmwgfx: Create VMWare host messaging module
This patch adds capabilities for a VMWare guest to send and
receive messages from the host, and adds functions to sending log
messages to vmware.log and to request device settings that aren't
available through the virtual hardware, e.g. certain settings in
the VMX file.

v2:
Created 32-bit compatible version of the macros.  This is much easily
done if "bp" is not also an output parameter, so modified both 32/64
versions to make "bp" an input-only parameter since currently there is
no need for it to be an output.

Signed-off-by: Sinclair Yeh <syeh@vmware.com>
Reviewed-by: Thomas Hellstrom <thellstrom@vmware.com>