vmwgfx:  Make the code build again on RHEL 7.x
RedHat, Ubuntu, and potentially other distros can selectively back port new
capabilities into an older kernel.  This means we can no longer rely on
LINUX_VERSION_CODE to provide the definitive "in/out" answer for some
of these compatibility code.

RHEL provides RHEL_MAJOR and RHEL_MINOR which helps a bit.  RHEL_MINOR in
this case really refers to the kernel, not the RHEL version.  For example,
a freshly installed 7.0 using the latest ISO or one that's been yum
updated will also use the same kernel as 7.3, i.e. RHEL_MINOR will be 3.

Ubuntu is in a similar situation.  dma_pool_zalloc() got backported into
4.2.0-42.  So on an un-updated Ubuntu 14.04 with 4.2.0-27, this new code
won't build.

From now on we will make the assumption that we will only build this module
on a fully updated distro.

* Changed rpc_msg_type -> vmwgfx_rpc_msg_type because the enum name conflicts
with another name on RHEL.

Signed-off-by: Sinclair Yeh <syeh@vmware.com>
Reviewed-by: Thomas Hellstrom <thellstrom@vmware.com>