vmwgfx: Fix screen target pinning and display surface replacement
Rework the code to unconditionally pin the display surface in
prepare_fp() and unpin in cleanup_fb(). Strictly we don't need the pinned
plane-state member anymore, but keep it for a while for debugging purposes.

Also make sure to allocate a new display surface on format- and dimension
mismatches. This caused device errors in some configurations.

Finally to avoid a clash between page-flip pinning and setcrtc pinning, modify
the page-flip code to use the page-flip helper and the atomic callbacks.
To enable this, we need to add a wrapper around atomic_commit.

Testing done: Fedora 25 operation with both Wayland and Xorg.

Signed-off-by: Thomas Hellstrom <thellstrom@vmware.com>
Reviewed-by: Sinclair Yeh <syeh@vmware.com>