vmwgfx/drm: Add core support to supply legacy cursor hotspot pointers
We revisit this change. Previous we added the different hotspots together
just before programming the device, but that makes the plane atomic state
incompatible with what's actually in the device.

So add a way for the only legacy ioctl that doesn't support cursor hotspots
to add the device internal legacy hotspots into the atomic plane state.

Signed-off-by: Thomas Hellstrom <thellstrom@vmware.com>
Reviewed-by: Brian Paul <brianp@vmware.com>