vmwgfx: Move texture max limit on topology to layout ioctl handler
Texture max limit on topology is to address some window managers that
create a big framebuffer for whole topology, so move it update layout
ioctl where the topology change request is actually received.

If this limit is removed then user-space fails either during surface
creation or at swrast depending on if it's a 3D VM or not.

Signed-off-by: Deepak Rawat <drawat@vmware.com>
Reviewed-by: Thomas Hellstrom <thellstrom@vmware.com>